Moved ClickableLabel to utils
