+++ src/Utils/ClickableLabel.py
@@ -1,0 +1,13 @@
+from PyQt5.QtGui import *
+from PyQt5.QtWidgets import *
+from PyQt5.QtCore import *
+from PyQt5.QtMultimedia import *
+from PyQt5.QtMultimediaWidgets import *
+
+class ClickableLabel(QLabel):
+    clicked = pyqtSignal()
+    def __init__(self, parent = None):
+        QLabel.__init__(self, parent)
+
+    def mousePressEvent(self, ev):
+        self.clicked.emit()
